Prepare all datasets and a to do list of models

Prepare all datasets and a to do list of models

prepare_analysis(
  repo,
  base,
  project,
  overwrite = FALSE,
  seed = 20070315,
  verbose = TRUE,
  min_observation = 100,
  min_stratum = 3,
  min_cycle = 2,
  proportion = 0.15,
  scheme_id = "ABV",
  volume,
  docker = "inbobmk/rn2k:0.7",
  dependencies = c("inbo/n2khelper@v0.4.3", "inbo/n2kanalysis@v0.2.9")
)

Arguments

repo

a git2rdata::repository() object

base

The base location to store the model.

project

Will be a relative path within the base location.

overwrite

Should an existing object be overwritten? Defaults to TRUE.

seed

The seed for the analysis. Defaults to 20070315.

verbose

display the progress

min_observation

The minimum number of observations for taking a species into account. Defaults to 100.

min_stratum

The required minimum number of relevant locations per stratum. Defaults to 3.

min_cycle

Minimum number of cycles in which a species must be found at a square. Square with sightings from fewer cycles will be ignored. Defaults to 2.

proportion

Minimum proportion of the least important period. Defaults to 0.15.

scheme_id

a string with the scheme id.

volume

The argument passed to the '-v' of docker. Only used when base is a local file system.

docker

The docker image to use during model fit.

dependencies

A vector of R packages as understood by the repo argument of remotes::install_github().